refactor: 콘솔 입력받는 클래스 object로 정의하는 방식으로 변경하고 콘솔 출력하는 클래스 구현 콘솔 입력받는 클래스 이름 또한 가독성 있게 변경함(GameConsoleInput->GameInputView)
